requirements.txt
pyyaml
click
rich
pydantic
py-imessage
pynput
